添加配置文件`

+++ src/index.ts
@@ -1,23 +1,31 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
 import walk from 'walk-sync';
+import stripJsonComments from 'strip-json-comments';
 const officegen = require('officegen');
+const config = JSON.parse(stripJsonComments(fs.readFileSync('src/config.json','utf-8')));
 
-const root = "C:/Users/pc/Desktop/code";//提取路径
-const globs = ['src/**/*.ts','src/**/*.tsx'];//提取范围
+
+const root = config.project;//提取路径
+const globs = config.globs;//提取范围
 const temp = 'out/temp.txt';
-const excludeFile:string[] = [];//排除的文件，文件名
-const excludeLine:string[] = [];//排除的行，字符串包含
+const excludeFileName:string[] = config.excludeFileName;//排除的文件，文件名
+const excludeFilePath:string[] = config.excludeFilePath;//排除的行，字符串包含
+const excludeLineText:string[] = config.excludeLineText;//排除的行，字符串包含
 const headerText = 'XXXXV1.0  源代码';
+const fontFace = config.fontFace;
+const fontSize = config.fontSize;
 async function run() {
+    if(!fs.existsSync(root))return;
     let paths = await getAllFiles(root,globs);
-    paths = excludeFiles(paths,excludeFile);
+    paths = filterFileName(paths,excludeFileName);
+    paths = filterFilePath(paths,excludeFilePath);
     fs.removeSync(temp);
     fs.ensureFileSync(temp);
     let lines:string[] = [];
     for (const e of paths) {
-        const ct = formatFile2Content(path.join(root,e));
-        lines = lines.concat(filterLines(ct.split('\n'),excludeLine));
+        const ct = formatFile2Content(e);
+        lines = lines.concat(filterLines(ct.split('\n'),excludeLineText));
         fs.appendFileSync(temp,ct);
 
     }
@@ -29,12 +37,16 @@ export function filterLines(src:string[],ext:string[]){
     return src.filter((e)=>ext.every((ee)=>!e.includes(ee)));
 }
 ///按文件名排除
-function excludeFiles(paths:string[],exc:string[]){
+export function filterFileName(paths:string[],exc:string[]){
     return paths.filter((e)=>exc.indexOf(path.basename(e))===-1);
+}
+///按文件路径排除
+export function filterFilePath(paths:string[],exc:string[]){
+    return paths.filter((e)=>exc.indexOf(e)===-1);
 }
 ///获取文件列表
 async function getAllFiles(root:string,globs:string[]){
-    const paths = walk(root,{ globs: globs}); 
+    const paths = walk(root,{ globs: globs}).map((e)=>path.join(root,e)).map((e)=>e.split(path.sep).join('/')); 
     return paths;
 }
 ///读取文件并返回过滤空行和注释的内容
@@ -55,9 +67,7 @@ function replaceNote(src:string){
 ///生成doc
 function genDoc(src:string){
     const docx = officegen({
-        type:'docx',
-        author:'leeyang1990@gmail.com',
-        creator:'leeyang1990@gmail.com',
+        type:'docx'
     });
     docx.on('finalize', function(written: any) {
         console.log(
@@ -68,7 +78,7 @@ function genDoc(src:string){
         console.log(err);
     });
     const pObj = docx.createP();
-    pObj.addText(src,{ font_face: 'Arial', font_size: 10 });
+    pObj.addText(src,{ font_face: fontFace, font_size: fontSize });
 
     const header = docx.getHeader().createP();
     header.addText ( headerText);
